rewrite derived table to use `Slot` (LRU not yet supported)
